helper move to directory
